--- examples/dev-server-no-compress/.umirc.ts
@@ -1,7 +1,0 @@
-export default {
-  proxy: {
-    '/events/': {
-      target: 'http://localhost:3000',
-    },
-  },
-};
