+++ examples/with-no-compress-for-sse/.umirc.ts
@@ -1,0 +1,1 @@
+export default {};
