+++ examples/with-no-compress-for-sse/package.json
@@ -1,11 +1,10 @@
 {
-  "name": "@example/dev-server-no-compress",
+  "name": "@example/with-no-compress-for-sse",
   "private": true,
   "scripts": {
     "build": "umi build",
     "dev": "umi dev",
-    "nocompress": "cross-env UMI_DEV_SERVER_COMPRESS=none npm run dev",
-    "withcompress": "cross-env npm run dev"
+    "dev:nocompress": "cross-env UMI_DEV_SERVER_COMPRESS=none npm run dev"
   },
   "dependencies": {
     "umi": "workspace:*"
