+++ examples/with-no-compress-for-sse/plugin.ts
@@ -1,0 +1,8 @@
+import { IApi } from 'umi';
+import { sseMiddleware } from './sse-middleware';
+
+export default (api: IApi) => {
+  api.onBeforeMiddleware(({ app }) => {
+    sseMiddleware(app);
+  });
+};
